implemented TaskListTest
